Put LRUCachePolicy in its own file.
