+++ src/main/scala/shark/memstore2/CachePolicy.scala
@@ -17,9 +17,6 @@
 
 package shark.memstore2
 
-import com.google.common.cache._
-import scala.collection.JavaConversions._
-
 
 trait CachePolicy[K, V] {
 
@@ -53,73 +50,4 @@ trait CachePolicy[K, V] {
   def getHitRate: Option[Double] = None
 
   def getEvictionCount: Option[Long] = None
-}
-
-class LRUCachePolicy[K <: AnyRef, V <: AnyRef] extends CachePolicy[K, V] {
-
-  var isInitialized = false
-  var hasRecordedStats = false
-  var cache: LoadingCache[K, V] = _
-  var cacheStats: Option[CacheStats] = None
-
-  override def initialize(
-	    maxSize: Long,
-	    loadFunc: (K => V),
-	    evictionFunc: (K, V) => Unit,
-	    shouldRecordStats: Boolean
-    ): Unit = {
-    super.initialize(maxSize, loadFunc, evictionFunc, shouldRecordStats)
-
-    var builder = CacheBuilder.newBuilder().maximumSize(maxSize)
-    if (shouldRecordStats) {
-      builder.recordStats()
-      hasRecordedStats = true
-    }
-
-    val removalListener =
-      new RemovalListener[K, V] {
-        def onRemoval(removal: RemovalNotification[K, V]): Unit = {
-          evictionFunc(removal.getKey, removal.getValue)
-        }
-      }
-    val cacheLoader =
-      new CacheLoader[K, V] {
-        def load(key: K): V = loadFunc(key)
-      }
-
-    cache = builder
-      .removalListener(removalListener)
-      .build(cacheLoader)
-    isInitialized = true
-  }
-
-  override def notifyPut(key: K, value: V): Unit = {
-    assert(isInitialized, "Must initialize() LRUCachePolicy.")
-    cache.put(key, value)
-  }
-
-  override def notifyRemove(key: K, value: V): Unit = {
-    assert(isInitialized, "Must initialize() LRUCachePolicy.")
-    cache.invalidate(key, value)
-  }
-
-  override def notifyGet(key: K): Unit = {
-    assert(isInitialized, "Must initialize() LRUCachePolicy.")
-    cache.get(key)
-  }
-  
-  override def getKeysOfCachedEntries: Seq[K] = {
-    assert(isInitialized, "Must initialize() LRUCachePolicy.")
-    return cache.asMap.keySet.toSeq
-  }
-
-  override def getHitRate(): Option[Double] = {
-    val hitRate = if (hasRecordedStats) Some(cache.stats.hitRate) else None
-    return hitRate
-  }
-
-  override def getEvictionCount(): Option[Long] = {
-    val evictionCount = if (hasRecordedStats) Some(cache.stats.evictionCount) else None
-    return evictionCount
-  }
 }
